requirements.txt
beautifulsoup4
imap-tools
requests
python-dotenv